Support multi-file grant structure
Some people share multiple documents making up a single grant separately and
some share reviews and other responses to the grants. This allows those
multiple documents to be separately linked to and labeled.

For UI this requires switching from a direct link to the proposal from the main
table to a link to a proposal page that includes one or more links to the proposal.